implemented basic caching
